Navigation Bar Hidden
Navigation Bar is hidden now and the game looks sleek and much better.
Yes we need few more redesigns and layout changes for some views.

I have not removed the navigation controller as it helps to sequence
easily so as you said Mam this is the finest way to remove the
Navigation controller. A small fix to hide the navigation bar.